refactor: Simplify report generation
