+++ src/qibocal/web/report.py
@@ -1,29 +1,36 @@
-import os
 import pathlib
+from dataclasses import dataclass
+from typing import Callable
 
-from jinja2 import Environment, FileSystemLoader
-
-from qibocal import __version__
-from qibocal.cli.report import ReportBuilder
+from qibocal.auto.history import History
+from qibocal.auto.task import TaskId
 
 WEB_DIR = pathlib.Path(__file__).parent
 STYLES = WEB_DIR / "static" / "styles.css"
 TEMPLATES = WEB_DIR / "templates"
 
 
-def create_report(path, report: ReportBuilder):
-    """Creates an HTML report for the data in the given path."""
-    with open(STYLES) as file:
-        css_styles = f"<style>\n{file.read()}\n</style>"
+@dataclass
+class Report:
+    """Report generation class."""
 
-    env = Environment(loader=FileSystemLoader(TEMPLATES))
-    template = env.get_template("template.html")
-    html = template.render(
-        is_static=True,
-        css_styles=css_styles,
-        version=__version__,
-        report=report,
-    )
+    path: pathlib.Path
+    """Path with calibration data."""
+    targets: list
+    """Global targets."""
+    history: History
+    """History of protocols."""
+    meta: dict
+    """Meta data."""
+    plotter: Callable
+    """Plotting function to generate html."""
 
-    with open(os.path.join(path, "index.html"), "w") as file:
-        file.write(html)
+    def routine_name(self, routine, iteration):
+        """Prettify routine's name for report headers."""
+        name = routine.replace("_", " ").title()
+        return f"{name} - {iteration}"
+
+    def routine_targets(self, task_id: TaskId):
+        """Get local targets parameter from Task if available otherwise use global one."""
+        local_targets = self.history[task_id].task.targets
+        return local_targets if len(local_targets) > 0 else self.targets
